Refactor exit module
Will delete the database if it exists or return true if it already doesn't exist

+++ src/exit.js
@@ -2,7 +2,19 @@ var Promise = require("bluebird");
 var fs = require("fs");
 var db = require("./store/db");
 var unlink = Promise.promisify(fs.unlink);
+var readFile = Promise.promisify(fs.readFile);
 function exit() {
+    return databaseExists()
+        .then(teardown);
+}
+function databaseExists() {
+    return readFile("slownode.db")
+        .then(function () { return true; })
+        .catch(function () { return false; });
+}
+function teardown(exists) {
+    if (exists)
+        return Promise.resolve(true);
     return db.destroy()
         .then(function () { return unlink("slownode.db"); })
         .then(function () { return true; })
