+++ src/exit.ts
@@ -2,9 +2,23 @@ import Promise = require("bluebird")
 import fs = require("fs");
 import db = require("./store/db");
 var unlink = Promise.promisify(fs.unlink);
+var readFile = Promise.promisify(fs.readFile);
 export = exit;
 
 function exit() {
+	return databaseExists()
+		.then(teardown);
+}
+
+function databaseExists() {
+	return readFile("slownode.db")
+		.then(() => true)
+		.catch(() => false);
+}
+
+function teardown(exists: boolean) {
+	if (exists) return Promise.resolve(true);
+
 	return db.destroy()
 		.then(() => unlink("slownode.db"))
 		.then(() => true)
